// src/types/invidious.ts
export interface ThumbnailObject {
  quality: string;
  url: string;
  width: number;
  height: number;
}
// ...
export interface PlaylistVideoObject {
  title: string;
  videoId: string;
  author?: string;
  authorId?: string;
  authorUrl?: string;
  lengthSeconds?: number;
  index?: number;
  indexId?: string;
  videoThumbnails: ThumbnailObject[];
}

export interface PlaylistObject {
  type: "playlist" | "invidiousPlaylist";
  title: string;
  playlistId: string;
  playlistThumbnail?: string;
  author: string;
  authorId?: string | null;
  authorUrl?: string | null;
  authorVerified?: boolean;
  videoCount?: number;
  videos?: PlaylistVideoObject[];
}